Streamlit Viz/streamlit/requirements.txt
altair
branca
matplotlib
numpy
pandas
pydeck
requests
typing
snowflake-snowpark-python
streamlit